+++ src/Plugin.php
@@ -4,16 +4,21 @@ namespace born05\twofactorauthentication;
 use born05\twofactorauthentication\widgets\Notify as NotifyWidget;
 
 use Craft;
+use born05\twofactorauthentication\services\Request as RequestService;
+use born05\twofactorauthentication\services\Response as ResponseService;
+use born05\twofactorauthentication\services\Verify as VerifyService;
+use born05\twofactorauthentication\models\Settings;
+use born05\twofactorauthentication\Variables;
 use craft\base\Plugin as CraftPlugin;
 use craft\base\Element;
-// use craft\elements\User;
+use craft\elements\User as UserElement;
 use craft\services\Dashboard;
 use craft\events\RegisterUrlRulesEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterElementTableAttributesEvent;
 use craft\events\SetElementTableAttributeHtmlEvent;
-use craft\helpers\UrlHelper;
 use craft\web\UrlManager;
+use craft\web\twig\variables\CraftVariable;
 use yii\base\Event;
 use yii\web\UserEvent;
 use born05\twofactorauthentication\web\User;
@@ -24,7 +29,7 @@ class Plugin extends CraftPlugin
      * @var string
      */
     public $schemaVersion = '2.0.0';
-    
+
     /**
      * @var bool
      */
@@ -43,55 +48,62 @@ class Plugin extends CraftPlugin
         parent::init();
         self::$plugin = $this;
 
-        if (!$this->isInstalled) return;
+        if (!$this->isInstalled) {
+            return;
+        }
+
+        $request = Craft::$app->getRequest();
+        $response = Craft::$app->getResponse();
+
+        if (!$this->isInstalled || $request->getIsConsoleRequest()) {
+            return;
+        }
+
+        // Register Components (Services)
+        $this->setComponents([
+            'request' => RequestService::class,
+            'response' => ResponseService::class,
+            'verify' => VerifyService::class,
+        ]);
+
+        // $this->request->validateRequest();
+
 
         // Verify after login.
-        Event::on(\craft\web\User::class, \craft\web\User::EVENT_AFTER_LOGIN, function(UserEvent $event) {
-            $user = Craft::$app->getUser()->getIdentity();
-            $request = Craft::$app->getRequest();
-            $response = Craft::$app->getResponse();
-
-            if (isset($user) &&
-                Plugin::$plugin->verify->isEnabled($user) &&
-                !Plugin::$plugin->verify->isVerified($user)
-            ) {
-                $url = UrlHelper::actionUrl('two-factor-authentication/verify/login');
-
-                if ($request->getAcceptsJson()) {
-                    return Plugin::$plugin->response->asJson(array(
-                        'success' => true,
-                        'returnUrl' => $url
-                    ));
-                } else {
-                    $response->redirect($url);
-                }
-            }
+        Event::on(\craft\web\User::class, \craft\web\User::EVENT_AFTER_LOGIN, function (UserEvent $event) {
+            $this->request->userLoginEventHandler($event);
         });
-        
-        Event::on(UrlManager::class, UrlManager::EVENT_REGISTER_CP_URL_RULES, function(RegisterUrlRulesEvent $event) {
-            $event->rules['two-factor-authentication'] = 'two-factor-authentication/default/index';
+
+        Event::on(CraftVariable::class, CraftVariable::EVENT_INIT, function (Event $event) {
+            /** @var CraftVariable $variable */
+            $variable = $event->sender;
+            $variable->set('twoFactorAuthentication', Variables::class);
         });
-        
+
+        Event::on(UrlManager::class, UrlManager::EVENT_REGISTER_CP_URL_RULES, function (RegisterUrlRulesEvent $event) {
+            $event->rules['two-factor-authentication'] = 'two-factor-authentication/settings/index';
+        });
+
         // Register our widgets
         Event::on(Dashboard::class, Dashboard::EVENT_REGISTER_WIDGET_TYPES, function (RegisterComponentTypesEvent $event) {
             $event->types[] = NotifyWidget::class;
         });
-        
+
         /**
          * Adds the following attributes to the User table in the CMS
          * NOTE: You still need to select them with the 'gear'
          */
-        Event::on(\craft\elements\User::class, Element::EVENT_REGISTER_TABLE_ATTRIBUTES, function(RegisterElementTableAttributesEvent $event) {
+        Event::on(UserElement::class, Element::EVENT_REGISTER_TABLE_ATTRIBUTES, function (RegisterElementTableAttributesEvent $event) {
             $event->tableAttributes['hasTwoFactorAuthentication'] = ['label' => Craft::t('two-factor-authentication', '2-Factor Auth')];
         });
 
         /**
          * Returns the content for the additional attributes field
          */
-        Event::on(\craft\elements\User::class, Element::EVENT_SET_TABLE_ATTRIBUTE_HTML, function(SetElementTableAttributeHtmlEvent $event) {
+        Event::on(UserElement::class, Element::EVENT_SET_TABLE_ATTRIBUTE_HTML, function (SetElementTableAttributeHtmlEvent $event) {
             $currentUser = Craft::$app->getUser()->getIdentity();
             if ($event->attribute == 'hasTwoFactorAuthentication' && $currentUser->admin) {
-                /** @var UserModel $user */
+                /** @var User $user */
                 $user = $event->sender;
 
                 if (Plugin::$plugin->verify->isEnabled($user)) {
@@ -104,6 +116,26 @@ class Plugin extends CraftPlugin
                 $event->handled = true;
             }
         });
+        
+        /**
+         * Hook into the users cp page.
+         */
+        Craft::$app->view->hook('cp.users.edit.details', function (array &$context) {
+            if (Craft::$app->getUser()->getIsAdmin() && !$context['isNewUser']) {
+                /** @var User $user */
+                $user = $context['user'];
+
+                return Craft::$app->getView()->renderTemplate('two-factor-authentication/_user/status', [
+                    'user' => $user,
+                    'enabled' => Plugin::$plugin->verify->isEnabled($user),
+                ]);
+            }
+        });
+    }
+
+    protected function createSettingsModel()
+    {
+        return new Settings();
     }
     
     /**