+++ src/controllers/SettingsController.php
@@ -7,23 +7,49 @@ use craft\web\Controller;
 use craft\elements\User;
 use craft\helpers\UrlHelper;
 use born05\twofactorauthentication\Plugin as TwoFactorAuth;
+use born05\twofactorauthentication\web\assets\verify\VerifyAsset;
 
 class SettingsController extends Controller
 {
     /**
+     * Show the settings form.
+     */
+    public function actionIndex()
+    {
+        Craft::$app->view->registerAssetBundle(VerifyAsset::class);
+        return $this->renderTemplate('two-factor-authentication/index');
+    }
+    
+    /**
+     * Show the settings form.
+     */
+    public function actionForce()
+    {
+        $user = Craft::$app->getUser()->getIdentity();
+
+        if (TwoFactorAuth::$plugin->verify->isVerified($user)) {
+            return Craft::$app->getResponse()->redirect(UrlHelper::cpUrl());
+        }
+
+        Craft::$app->view->registerAssetBundle(VerifyAsset::class);
+        return $this->renderTemplate('two-factor-authentication/_force');
+    }
+
+    /**
      * Turn on 2-factor for current user.
      */
-    public function actionTurnOn() {
+    public function actionTurnOn()
+    {
         $this->requirePostRequest();
 
         $user = Craft::$app->getUser()->getIdentity();
-        $requestService = Craft::$app->getRequest();
-        
-        $authenticationCode = $requestService->getBodyParam('authenticationCode');
-        $returnUrl = UrlHelper::cpUrl('two-factor-authentication');
+        $request = Craft::$app->getRequest();
+  
+        $authenticationCode = $request->getBodyParam('authenticationCode');
+        $returnUrl = TwoFactorAuth::$plugin->response->getReturnUrl();
 
         if (TwoFactorAuth::$plugin->verify->verify($user, $authenticationCode)) {
-            if ($requestService->getAcceptsJson()) {
+            if ($request->getAcceptsJson()) {
                 return $this->asJson([
                     'success' => true,
                     'returnUrl' => $returnUrl
@@ -35,15 +61,15 @@ class SettingsController extends Controller
             $errorCode = User::AUTH_INVALID_CREDENTIALS;
             $errorMessage = Craft::t('two-factor-authentication', 'Authentication code is invalid.');
 
-            if ($requestService->getAcceptsJson()) {
+            if ($request->getAcceptsJson()) {
                 return $this->asJson([
                     'errorCode' => $errorCode,
                     'error' => $errorMessage
                 ]);
             } else {
-                Craft::$app->user->setError($errorMessage);
+                Craft::$app->getSession()->setError($errorMessage);
 
-                Craft::$app->urlManager->setRouteVariables([
+                Craft::$app->getUrlManager()->setRouteParams([
                     'errorCode' => $errorCode,
                     'errorMessage' => $errorMessage,
                 ]);
@@ -55,7 +81,8 @@ class SettingsController extends Controller
     /**
      * Disable 2-factor for current user.
      */
-    public function actionTurnOff() {
+    public function actionTurnOff()
+    {
         $this->requirePostRequest();
 
         $user = Craft::$app->getUser()->getIdentity();