+++ src/migrations/Install.php
@@ -3,6 +3,7 @@ namespace born05\twofactorauthentication\migrations;
 
 use craft\db\Migration;
 use craft\db\Query;
+use craft\helpers\MigrationHelper;
 use born05\twofactorauthentication\records\Session;
 use born05\twofactorauthentication\records\User;
 
@@ -10,10 +11,10 @@ class Install extends Migration
 {
     public function safeUp()
     {
-        if ($this->_upgradeFromCraft2()) {
+        if ($this->upgradeFromCraft2()) {
             return;
         }
-        
+
         // Fresh install code goes here...
         $this->createTable(Session::tableName(), [
             'id' => $this->primaryKey(),
@@ -24,7 +25,7 @@ class Install extends Migration
             'dateUpdated' => $this->dateTime()->notNull(),
             'uid' => $this->uid(),
         ]);
-        
+
         $this->createTable(User::tableName(), [
             'id' => $this->primaryKey(),
             'userId' => $this->integer()->notNull(),
@@ -34,37 +35,26 @@ class Install extends Migration
             'dateUpdated' => $this->dateTime()->notNull(),
             'uid' => $this->uid(),
         ]);
-        
+
         $this->createIndex(null, Session::tableName(), ['userId', 'sessionId'], true);
         $this->createIndex(null, User::tableName(), ['userId'], true);
-        
+
         $this->addForeignKey(null, Session::tableName(), ['userId'], '{{%users}}', ['id'], 'CASCADE', null);
         $this->addForeignKey(null, Session::tableName(), ['sessionId'], '{{%sessions}}', ['id'], 'CASCADE', null);
         $this->addForeignKey(null, User::tableName(), ['userId'], '{{%users}}', ['id'], 'CASCADE', null);
     }
 
-    private function _upgradeFromCraft2()
+    protected function upgradeFromCraft2()
     {
         // Fetch the old plugin row, if it was installed
-        $row = (new Query())
-            ->select(['id', 'settings'])
-            ->from(['{{%plugins}}'])
-            ->where(['in', 'handle', ['twofactorauthentication']])
-            ->one();
-        
-        if (!$row) {
+        $oldTableExists = $this->db->schema->getTableSchema('{{%twofactorauthentication_sessions}}', true);
+        if (!isset($oldTableExists)) {
             return false;
         }
 
-        // Update this one's settings to old values
-        $this->update('{{%plugins}}', [
-            'settings' => $row['settings']
-        ], ['handle' => 'two-factor-authentication']);
-
-        // Delete the old row
-        $this->delete('{{%plugins}}', ['id' => $row['id']]);
-
         // Any additional upgrade code goes here...
+        $this->renameTable('{{%twofactorauthentication_sessions}}', Session::tableName());
+        $this->renameTable('{{%twofactorauthentication_users}}', User::tableName());
 
         return true;
     }